SUIT encryption: tag placed at the start of the firmware
This allows for a much easier documentation on the device side.
The streaming version of PSA API needs to pass the tag separately.
It makes it much easier to identify the tag if it is placed at
the start of the encrypted payload, as it doesn't require to know
the whole payload size in advanced.

At the same time no scenario is planned in which the non-streaming
version of psa decryption API is used by SUIT, so placing the
tag at the end of the payload is not required.

Ref: NCSDK-28265

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>